+++ packages/opentelemetry-resources/test/index-webpack.worker.ts
@@ -14,6 +14,12 @@
  * limitations under the License.
  */
 
-export * from './BrowserDetector';
-export * from './EnvDetector';
-export * from './ProcessDetector';
+{
+  const testsContext = require.context('./', false, /test$/);
+  testsContext.keys().forEach(testsContext);
+}
+
+{
+  const testsContext = require.context('./detectors/browser', false, /test$/);
+  testsContext.keys().forEach(testsContext);
+}